# docs/openapi.yaml
openapi: 3.0.3
info:
  title: CoopLumen API
  description: REST API for CoopLumen communities and Stellar assets.
  version: 0.1.0
servers:
  - url: http://localhost:4000
...
paths:
...
  /api/v1/tokens/transfer:
    post:
      tags: [Transactions]
      summary: Submit a signed Stellar payment transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenTransferRequest'
      responses:
        '200':
          description: Payment submitted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenTransferResponse'
        '400': { $ref: '#/components/responses/ValidationError' }
        '402':
          description: The source account does not have enough XLM for the payment and fee
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InsufficientBalanceResponse'
        '422':
          description: The Stellar network rejected the signed transaction
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Horizon is temporarily unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/tokens/airdrop:
    post:
      tags: [Transactions]
      summary: Airdrop a community token to all community members
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AirdropRequest'
      responses:
        '200':
          description: Airdrop submitted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AirdropResponse'
        '400': { $ref: '#/components/responses/ValidationError' }
        '402':
          description: The issuer account does not have enough XLM for transaction fees
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InsufficientBalanceResponse'
        '404': { $ref: '#/components/responses/NotFound' }
        '422':
          description: The Stellar network rejected the airdrop transaction
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Horizon is temporarily unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/tokens/{assetCode}/{issuer}:
    get:
      tags: [Tokens]
      summary: Get token metadata
      description: Returns the off-chain metadata for one Stellar custom asset.
      parameters:
        - name: assetCode
          in: path
          required: true
          description: Stellar asset code.
          schema:
            type: string
            pattern: '^[A-Za-z0-9]{1,12}$'
        - name: issuer
          in: path
          required: true
          description: Stellar public key of the asset issuer.
          schema:
            type: string
            pattern: '^G[A-Z2-7]{55}$'
      responses:
        '200':
          description: Token metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenMetadataResponse'
        '400': { $ref: '#/components/responses/ValidationError' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/tokens/{communityId}:
    get:
      tags: [Tokens]
      summary: List tokens issued for a community
      parameters:
        - name: communityId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Tokens issued for the community
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Token' }
  /api/v1/tokens/burn:
    post:
      tags: [Tokens]
      summary: Burn tokens
      description: Reduces circulating supply by returning tokens to the issuing account.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BurnToken' }
      responses:
        '200':
          description: Tokens burned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenTransferResponse'
        '400': { $ref: '#/components/responses/ValidationError' }
        '502':
          description: Horizon is temporarily unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/tokens/trustline:
    post:
      tags: [Tokens]
      summary: Establish a trustline
      description: Allows an account to hold a community token; must be called before it can receive the asset.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [accountSecret, assetCode, assetIssuer]
              properties:
                accountSecret: { type: string }
                assetCode: { type: string }
                assetIssuer: { $ref: '#/components/schemas/StellarKey' }
                limit: { type: string }
      responses:
        '201':
          description: Trustline established
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenTransferResponse'
        '400': { $ref: '#/components/responses/ValidationError' }
        '502':
          description: Horizon is temporarily unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/tokens/holders/{assetCode}/{issuer}:
    parameters:
      - name: assetCode
        in: path
        required: true
        schema: { type: string, pattern: '^[A-Za-z0-9]{1,12}$' }
      - name: issuer
        in: path
        required: true
        schema: { $ref: '#/components/schemas/StellarKey' }
    get:
      tags: [Tokens]
      summary: List accounts holding a token
      description: Backed by Horizon's asset endpoint.
      responses:
        '200':
          description: Accounts holding the asset
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        account: { $ref: '#/components/schemas/StellarKey' }
                        balance: { type: string }
        '400': { $ref: '#/components/responses/ValidationError' }
        '502':
          description: Horizon is temporarily unavailable
  /api/v1/tokens/supply/{assetCode}/{issuer}:
    parameters:
      - name: assetCode
        in: path
        required: true
        schema: { type: string, pattern: '^[A-Za-z0-9]{1,12}$' }
      - name: issuer
        in: path
        required: true
        schema: { $ref: '#/components/schemas/StellarKey' }
    get:
      tags: [Tokens]
      summary: Get circulating supply
      description: Returns the circulating supply Horizon reports for the asset.
      responses:
        '200':
          description: Circulating supply
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      assetCode: { type: string }
                      issuer: { $ref: '#/components/schemas/StellarKey' }
                      supply: { type: string }
        '400': { $ref: '#/components/responses/ValidationError' }
        '502':
          description: Horizon is temporarily unavailable
  /api/v1/tokens/history/{assetCode}/{issuer}:
    parameters:
      - name: assetCode
        in: path
        required: true
        schema: { type: string, pattern: '^[A-Za-z0-9]{1,12}$' }
      - name: issuer
        in: path
        required: true
        schema: { $ref: '#/components/schemas/StellarKey' }
    get:
      tags: [Tokens]
      summary: Get recent payment activity for a token
      description: >
        Horizon has no dedicated "payments for an asset" endpoint, so this
        proxies the issuer account's payment history filtered to the asset.
      responses:
        '200':
          description: Recent payment activity
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { type: object }
        '400': { $ref: '#/components/responses/ValidationError' }
...
  /api/v1/tokens/issue:
    post:
      tags: [Tokens]
      summary: Issue a community token
      description: >
        Issues a community token on the Stellar network. Accepts an optional
        Idempotency-Key header; a retried request with the same key replays
        the original response instead of issuing a second time.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/IssueToken' }
      responses:
        '201':
          description: Token issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenTransferResponse'
        '400': { $ref: '#/components/responses/ValidationError' }
        '429':
          description: Too many requests. Rate limited to 3 per minute per user.
        '500':
          description: The asset was issued but its metadata could not be persisted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Horizon is temporarily unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
...
  responses:
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
    NotFound:
      description: Requested resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    StellarKey:
      type: string
      description: A 56-character Stellar StrKey public key.
      pattern: '^G[A-Z2-7]{55}$'
...
    ValidationIssue:
      type: object
      properties:
        path:
          oneOf:
            - type: string
            - type: array
              items: { type: string }
        message: { type: string }
    ValidationErrorResponse:
      type: object
      required: [data, error]
      properties:
        data:
          nullable: true
        meta:
          type: object
          properties:
            errors:
              type: array
              items:
                $ref: '#/components/schemas/ValidationIssue'
        error:
          oneOf:
            - type: string
            - type: object
              additionalProperties: true
...
    TokenTransferRequest:
      type: object
      required: [signedXdr]
      properties:
        signedXdr:
          type: string
          minLength: 1
          maxLength: 100000
    TokenTransferResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          properties:
            txHash: { type: string }
...
    AirdropRequest:
      type: object
      required: [communityId, amount, issuerSecret]
      properties:
        communityId:
          type: string
          format: uuid
        amount:
          type: string
          description: Positive decimal amount with up to 7 fractional digits.
        issuerSecret:
          type: string
          description: Stellar secret key for the issuing account.
    AirdropResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          properties:
            amount: { type: string }
            recipientCount: { type: integer }
            txHashes:
              type: array
              items: { type: string }
    InsufficientBalanceError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: INSUFFICIENT_BALANCE
        message:
          type: string
        requiredXlm:
          type: string
          nullable: true
        currentBalance:
          type: string
          nullable: true
    InsufficientBalanceResponse:
      type: object
      required: [data, error]
      properties:
        data:
          nullable: true
        error:
          $ref: '#/components/schemas/InsufficientBalanceError'
    Token:
      type: object
      properties:
        id: { type: string, format: uuid }
        community_id: { type: string, format: uuid }
        asset_code: { type: string }
        asset_issuer: { $ref: '#/components/schemas/StellarKey' }
        distributor_address: { $ref: '#/components/schemas/StellarKey' }
        total_supply: { type: string }
        description: { type: string, nullable: true }
        icon_url: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    BurnToken:
      type: object
      required: [asset_code, asset_issuer]
      properties:
        id:
          type: string
        community_id:
          type: string
          format: uuid
        asset_code:
          type: string
        asset_issuer:
          type: string
        name:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    TokenMetadata:
      type: object
      properties:
        id: { type: string, format: uuid }
        asset_code: { type: string }
        asset_issuer: { $ref: '#/components/schemas/StellarKey' }
        name: { type: string }
        description: { type: string, nullable: true }
    TokenMetadataResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/TokenMetadata'
    IssueToken:
      type: object
      required: [issuerSecret, assetCode, distributorPublicKey, amount]
      properties:
        communityId: { type: string, format: uuid }
        issuerSecret: { type: string }
        assetCode: { type: string, pattern: '^[A-Za-z0-9]{1,12}$' }
        distributorPublicKey: { $ref: '#/components/schemas/StellarKey' }
        amount: { type: string }
        memo: { type: string, maxLength: 28 }
    ErrorResponse:
      type: object
      required: [data, error]
      properties:
        data:
          nullable: true
        meta:
          type: object
          additionalProperties: true
        error:
          oneOf:
            - type: string
            - type: object
              additionalProperties: true
...